Remove compiler.cpp, compiler.h
